Repair small bugs and change long to BigInteger

+++ src/SumatorInterface.java
@@ -1,8 +1,9 @@
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
 public interface SumatorInterface {
     void run(String file);
-    void createFiboCollectionAndSumIndex(List<Long> fibonacciNumbers,List<Long> fileNumbers);
-    Map<Integer,Long> generateSpecialFibonacciCollection(Long max);
+    void createFiboCollectionAndSumIndex(List<BigInteger> fibonacciNumbers, List<Long> fileNumbers);
+    Map<Integer,BigInteger> generateSpecialFibonacciCollection(BigInteger max);
 }
